requirements.txt
pandas
numpy
matplotlib
seaborn
sqlalchemy
jupyterlab
notebook
python-dotenv
transformers
sentence-transformers
torch   
pandera
scikit-learn
tqdm
nbstripout